Fixing various bugs
