+++ services/net/transcription/appsettings.json
@@ -16,7 +16,7 @@
     "ApiUrl": "http://api:8080",
     "TimeZone": "UTC",
     "Topics": "transcribe",
-    "FilePath": "/data/clip",
+    "FilePath": "/data",
     "AzureRegion": "canadacentral",
     "AzureCognitiveServicesKey": ""
   },
